Initialise repository
This is the first commit of my work on the nos kernel. This project is
purely an entertainment/education exercise, and is based very heavily on
the James Molloy's OS development tutorials, available here:

    http://www.jamesmolloy.co.uk/tutorial_html/

At the moment, the kernel is in a very bare state, offering paging,
limited ISR, IRQ and PIT support, and direct access to the VGA
framebuffer through a weak TTY abstraction.

Over time I will be adding more modules and functionality, with the
emphasis being on simple and understandable designs instead of features,
compatibility or even stability (this project is purely a technical
exercise and I have no intention of it being used for anything other
than vaguely probing the field of operating system design and
implementation).