scripts/tags: Improve source directory selection
This changes the usage of the tags scripts from:

     tags.sh <srcdir>

to:

     tags.sh <directory ...>

The benefit of this is that rather than specifying the top of the source
tree and searching for all C sources within it (including, perhaps in
unwanted locations, such as .git/ or .docs/), you can now run the script
with more precise granularity, such as $ ./scripts/tag.sh path/dir1/
path/dir2/.

The make TAGS target has been update accordingly. Additionally, an unnecessary
variable has been removed.