Top level Makefile improvements
This patch combines three sets of improvements to the top level Makefile
operation:

  1) Replace kernel make target with per-dir targets
  2) Makefile 'run' target does not result in compilation
  3) Make clean target is made more generic

The top level Makefile kernel target has been replaced with an array of
targets, one for each of the members of the array $(SOURCEDIRS). This
allow for directory specific building. New source directories should be added to
this array.

This nos target dependency has been removed from the run target, so that
executing 'make run' does not update the kernel image. From now on, you must
call compilation specifically, using 'make all run'. This decouples the kernel
sources and build state from the operation of the emulator.