Add 'todo' script
This patch adds a file scripts/todo.sh which searches for and displays all
'TODO:' and 'FIXME:' tags in a list of directories, passed as command line
arguments. Additionally, a 'make todo' target has been added which executes this
script with a list of all of the source directories. Addresses issue #12.

  #12: Add a 'todo' viewer script

  There should be a script to search through all of the source files and display
  a list of all mentions of 'TODO:' and 'FIXME:' tags, complete with file name
  and line number, for example:

    $ ./scripts/todo
    nos/main.c:109    FIXME   Undefined reference error.
    nos/tty/ports.h   TODO    Refactor enum declaration.