Bodge major page fault regression
For reasons unknown to me, the kernel became unbootable. The problem was
caused by the creation of the kernel heap. This would cause a page
fault at address KHEAP_START (0xC0000000). By trial-and-error, I found a
value that works, however this is not a suitable solution, and the
regression should be pursued further.