nos: Separate includes from sources
The nos project tree is restructured so that the majority of headers can
be found in an include/ directory, and its accompanying
sub-directories. For example, the following source layout:

    isr.c
    isr.h
    sched/sched.c
    sched/sched.h

has been transformed into:

    isr.c
    sched/sched.h
    include/isr.h
    include/sched/sched.h

Addresses issue #19.

    #19: Header and source files are not separated

    Source and include files should be located in separate directories,
    with a directory include/ containing sub-directories for
    headers. This means that compilation can be achieved by passing -I
    include/ to gcc.

This patch takes advantage of these changes to also introduce some
higher quality documentation in the project Makefile.